Julio2

+++ Julio/para_hacerxyz.py
@@ -9,7 +9,7 @@ import pandas as pd
 
 #df = pd.read_csv('dem_all_z18.txt', sep=';', decimal=',', header=0, names=['FID', 'pointid', 'grid_code', 'X', 'Y'])
 df = pd.read_csv('UTM_Z18_30m.txt', sep=';', decimal=',', header=0)
-#print(df.astype(float))
+df = df.astype(float)
 '''
 pd.to_numeric(df['FID']) # Convertir de texto a float la columna
 pd.to_numeric(df['pointid'])
